Implement cursor hover effect on jewellery page
Each script has been given it's own
setupCustomCursorHoverEffect function that
implements the effect for the elements specific
to the html document it's linked to except for
main.js that implements the effect for elements
common to all documents (main.js is linked to
all html documents).

A cursor-shrinks css class enables a
shrinking effect on the cursor that maybe used
as desired similar to previous cursor-grows class.

Further, css transform properties have been
abandoned for centering the custom cursor
(relative to the original cursor) in favor of the
more obvious and simpler adjustment of it's
fixed positioning.